Move `FilmsFirebaseSource` to new package `remote`
